organize panelView into a directory
