+++ app/frontend/src/classes/PanelView/PanelView.js
