+++ app/frontend/src/classes/PanelView/PanelViewPreviewAlternativeAlleleFrequencies.js
@@ -1,6 +1,6 @@
 import PanelView from './PanelView.js';
-import { storeManager } from '../store/StoreManager';
-import { getSimpleSearchConditionMaster } from '../store/searchManager';
+import { storeManager } from '../../store/StoreManager';
+import { getSimpleSearchConditionMaster } from '../../store/searchManager';
 
 const DECIMAL_DIGIT = 4;
